+++ backend/internal/utils/utils.go
@@ -24,7 +24,7 @@ func CheckPasswordHash(password, hash string) bool {
 	return err == nil
 }
 
-func GenerateToken(userID uint, secret string) (string, error) {
+func GenerateToken(userID int64, secret string) (string, error) {
 	claims := jwt.MapClaims{
 		"user_id": userID,
 		"exp":     time.Now().Add(time.Hour * 72).Unix(),
